Make AlpacaToMessage public. (#1785)
