+++ torchtune/data/__init__.py
@@ -17,6 +17,7 @@ from torchtune.data._common import CROSS_ENTROPY_IGNORE_IDX
 from torchtune.data._converters import get_openai_messages, get_sharegpt_messages
 from torchtune.data._instruct_templates import InstructTemplate
 from torchtune.data._messages import (
+    AlpacaToMessages,
     ChosenRejectedToMessages,
     InputOutputToMessages,
     Message,
@@ -43,6 +44,7 @@ __all__ = [
     "SummarizeTemplate",
     "OpenAIToMessages",
     "ShareGPTToMessages",
+    "AlpacaToMessages",
     "truncate",
     "Message",
     "validate_messages",
